Move functions related test cases into test_functions.py
